+++ pom.xml
@@ -64,7 +64,7 @@
                     <execution>
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>org.example.pesadillamago/org.example.pesadillamago.HelloApplication</mainClass>
+                            <mainClass>org.example.pesadillamago/org.example.pesadillamago.ui.marcos.HelloApplication</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
