+++ src/main/java/module-info.java
@@ -6,4 +6,6 @@ module org.example.pesadillamago {
     exports org.example.pesadillamago;
     opens org.example.pesadillamago to javafx.fxml;
     opens org.example.pesadillamago.game.dungeon.home to javafx.fxml;
+    exports org.example.pesadillamago.ui.marcos;
+    opens org.example.pesadillamago.ui.marcos to javafx.fxml;
 }
