--- src/main/java/org/example/pesadillamago/HelloController.java
@@ -1,14 +1,0 @@
-package org.example.pesadillamago;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
