+++ src/main/java/org/example/pesadillamago/ui/marcos/HelloController.java
@@ -1,0 +1,14 @@
+package org.example.pesadillamago.ui.marcos;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class HelloController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
